teste/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.7
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
djangorestframework==3.14.0
grpcio==1.49.1
grpcio-tools==1.49.1
protobuf==4.21.7
psycopg2-binary==2.9.5
python-dotenv==1.0.0
pytz==2022.7.1
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.4.0
